skill/lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
ask-sdk==1.15.0
ask-sdk-model==1.30.1
ask-sdk-dynamodb-persistence-adapter==1.15.0